+++ core/src/stored/append.cc
@@ -25,19 +25,17 @@
  * Append code for Storage daemon
  */
 
-#include "include/bareos.h"
+#include "stored/append.h"
 #include "stored/stored.h"
 #include "stored/acquire.h"
-#include "stored/append.h"
-#include "stored/device_control_record.h"
 #include "stored/fd_cmds.h"
+#include "stored/stored_globals.h"
 #include "stored/jcr_private.h"
 #include "stored/label.h"
 #include "stored/spool.h"
 #include "lib/bget_msg.h"
 #include "lib/edit.h"
 #include "include/jcr.h"
-#include "lib/bsock.h"
 #include "lib/berrno.h"
 #include "lib/berrno.h"
 
@@ -52,22 +50,127 @@ static char OK_replicate[] = "3000 OK replicate data\n";
 
 void PossibleIncompleteJob(JobControlRecord* jcr, int32_t last_file_index) {}
 
+
+ProcessedFileData::ProcessedFileData(DeviceRecord* record)
+    : volsessionid_(record->VolSessionId)
+    , volsessiontime_(record->VolSessionTime)
+    , fileindex_(record->FileIndex)
+    , stream_(record->Stream)
+    , data_len_(record->data_len)
+    , data_(record->data, record->data + record->data_len)
+{
+}
+
+DeviceRecord ProcessedFileData::GetData()
+{
+  DeviceRecord devicerecord{};
+  devicerecord.VolSessionId = volsessionid_;
+  devicerecord.VolSessionTime = volsessiontime_;
+  devicerecord.FileIndex = fileindex_;
+  devicerecord.Stream = stream_;
+  devicerecord.data_len = data_len_;
+  devicerecord.data = data_.data();
+
+  return devicerecord;
+}
+
+ProcessedFile::ProcessedFile(int32_t fileindex) : fileindex_(fileindex) {}
+
+void ProcessedFile::SendAttributesToDirector(JobControlRecord* jcr)
+{
+  for_each(attributes_.begin(), attributes_.end(),
+           [&jcr](ProcessedFileData& attribute) {
+             DeviceRecord devicerecord = attribute.GetData();
+             SendAttrsToDir(jcr, &devicerecord);
+           });
+}
+
+void ProcessedFile::AddAttribute(DeviceRecord* record)
+{
+  attributes_.emplace_back(ProcessedFileData(record));
+}
+
+bool IsAttribute(DeviceRecord* record)
+{
+  return record->maskedStream == STREAM_UNIX_ATTRIBUTES
+         || record->maskedStream == STREAM_UNIX_ATTRIBUTES_EX
+         || record->maskedStream == STREAM_RESTORE_OBJECT
+         || CryptoDigestStreamType(record->maskedStream) != CRYPTO_DIGEST_NONE;
+}
+
+static void UpdateFileList(JobControlRecord* jcr)
+{
+  Dmsg0(100, _("... update file list\n"));
+  jcr->impl->dcr->DirAskToUpdateFileList();
+}
+
+static void UpdateJobmediaRecord(JobControlRecord* jcr)
+{
+  Dmsg0(100, _("... create job media record\n"));
+  jcr->impl->dcr->DirCreateJobmediaRecord(false);
+}
+
+static void UpdateJobrecord(JobControlRecord* jcr)
+{
+  Dmsg2(100, _("... update job record: %llu bytes %lu files\n"), jcr->JobBytes,
+        jcr->JobFiles);
+  jcr->impl->dcr->DirAskToUpdateJobRecord();
+}
+
+void DoBackupCheckpoint(JobControlRecord* jcr)
+{
+  Jmsg(jcr, M_INFO, 0,
+       _("Checkpoint: Syncing current backup status to catalog\n"));
+  UpdateJobrecord(jcr);
+  UpdateFileList(jcr);
+  UpdateJobmediaRecord(jcr);
+  Jmsg(jcr, M_INFO, 0, _("Checkpoint completed\n"));
+}
+
+static time_t DoTimedCheckpoint(JobControlRecord* jcr,
+                                time_t checkpoint_time,
+                                time_t checkpoint_interval)
+{
+  const time_t now
+      = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now());
+  if (now > checkpoint_time) {
+    while (checkpoint_time <= now) { checkpoint_time += checkpoint_interval; }
+    Jmsg(jcr, M_INFO, 0,
+         _("Doing timed backup checkpoint. Next checkpoint in %d seconds\n"),
+         checkpoint_interval);
+    DoBackupCheckpoint(jcr);
+  }
+
+  return checkpoint_time;
+}
+
+static void SaveFullyProcessedFiles(JobControlRecord* jcr,
+                                    std::vector<ProcessedFile>& processed_files)
+{
+  if (!processed_files.empty()) {
+    for_each(
+        processed_files.begin(), processed_files.end(),
+        [&jcr](ProcessedFile& file) { file.SendAttributesToDirector(jcr); });
+    jcr->JobFiles = processed_files.back().GetFileIndex();
+    processed_files.clear();
+  }
+}
+
 // Append Data sent from File daemon
 bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
 {
   int32_t n, file_index, stream, last_file_index, job_elapsed;
   bool ok = true;
   char buf1[100];
-  DeviceControlRecord* dcr = jcr->impl->dcr;
   Device* dev;
   POOLMEM* rec_data;
   char ec[50];
 
-  if (!dcr) {
+  if (!jcr->impl->dcr) {
     Jmsg0(jcr, M_FATAL, 0, _("DeviceControlRecord is NULL!!!\n"));
     return false;
   }
-  dev = dcr->dev;
+  dev = jcr->impl->dcr->dev;
   if (!dev) {
     Jmsg0(jcr, M_FATAL, 0, _("Device is NULL!!!\n"));
     return false;
@@ -75,16 +178,23 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
 
   Dmsg1(100, "Start append data. res=%d\n", dev->NumReserved());
 
-  if (!bs->SetBufferSize(dcr->device_resource->max_network_buffer_size,
-                         BNET_SETBUF_WRITE)) {
+  if (!bs->SetBufferSize(
+          jcr->impl->dcr->device_resource->max_network_buffer_size,
+          BNET_SETBUF_WRITE)) {
     Jmsg0(jcr, M_FATAL, 0, _("Unable to set network buffer size.\n"));
-    goto bail_out;
+    jcr->setJobStatus(JS_ErrorTerminated);
+    return false;
   }
 
-  if (!AcquireDeviceForAppend(dcr)) { goto bail_out; }
+  if (!AcquireDeviceForAppend(jcr->impl->dcr)) {
+    jcr->setJobStatus(JS_ErrorTerminated);
+    return false;
+  }
 
-  if (GeneratePluginEvent(jcr, bSdEventSetupRecordTranslation, dcr) != bRC_OK) {
-    goto bail_out;
+  if (GeneratePluginEvent(jcr, bSdEventSetupRecordTranslation, jcr->impl->dcr)
+      != bRC_OK) {
+    jcr->setJobStatus(JS_ErrorTerminated);
+    return false;
   }
 
   jcr->sendJobStatus(JS_Running);
@@ -94,11 +204,15 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
   }
   Dmsg1(50, "Begin append device=%s\n", dev->print_name());
 
-  if (!BeginDataSpool(dcr)) { goto bail_out; }
+  if (!BeginDataSpool(jcr->impl->dcr)) {
+    jcr->setJobStatus(JS_ErrorTerminated);
+    return false;
+  }
 
   if (!BeginAttributeSpool(jcr)) {
-    DiscardDataSpool(dcr);
-    goto bail_out;
+    DiscardDataSpool(jcr->impl->dcr);
+    jcr->setJobStatus(JS_ErrorTerminated);
+    return false;
   }
 
   Dmsg0(100, "Just after AcquireDeviceForAppend\n");
@@ -107,7 +221,7 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
   }
 
   // Write Begin Session Record
-  if (!WriteSessionLabel(dcr, SOS_LABEL)) {
+  if (!WriteSessionLabel(jcr->impl->dcr, SOS_LABEL)) {
     Jmsg1(jcr, M_FATAL, 0, _("Write session label failed. ERR=%s\n"),
           dev->bstrerror());
     jcr->setJobStatus(JS_ErrorTerminated);
@@ -142,8 +256,19 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
    * file. 1. for the Attributes, 2. for the file data if any,
    * and 3. for the MD5 if any.
    */
-  dcr->VolFirstIndex = dcr->VolLastIndex = 0;
+  jcr->impl->dcr->VolFirstIndex = jcr->impl->dcr->VolLastIndex = 0;
   jcr->run_time = time(NULL); /* start counting time for rates */
+
+  auto now
+      = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now());
+  time_t next_checkpoint_time = now + me->checkpoint_interval;
+
+  std::vector<ProcessedFile> processed_files{};
+  int64_t current_volumeid = jcr->impl->dcr->VolMediaId;
+
+  ProcessedFile file_currently_processed;
+  uint32_t current_block_number = jcr->impl->dcr->block->BlockNumber;
+
   for (last_file_index = 0; ok && !jcr->IsJobCanceled();) {
     /*
      * Read Stream header from the daemon.
@@ -180,24 +305,28 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
      * An incomplete job can start the file_index at any number.
      * otherwise, it must start at 1.
      */
-    if (jcr->rerunning && file_index > 0 && last_file_index == 0) {
-      goto fi_checked;
-    }
-    if (file_index > 0
-        && (file_index == last_file_index
-            || file_index == last_file_index + 1)) {
-      goto fi_checked;
-    }
-    Jmsg3(jcr, M_FATAL, 0, _("FI=%d from %s not positive or sequential=%d\n"),
-          file_index, what, last_file_index);
-    PossibleIncompleteJob(jcr, last_file_index);
-    ok = false;
-    break;
 
-  fi_checked:
+    bool incomplete_job_rerun_fileindex_positive
+        = jcr->rerunning && file_index > 0 && last_file_index == 0;
+    bool fileindex_is_sequential = file_index > 0
+                                   && (file_index == last_file_index
+                                       || file_index == last_file_index + 1);
+
+    if (!incomplete_job_rerun_fileindex_positive && !fileindex_is_sequential) {
+      Jmsg3(jcr, M_FATAL, 0,
+            _("FileIndex=%d from %s not positive or sequential=%d\n"),
+            file_index, what, last_file_index);
+      PossibleIncompleteJob(jcr, last_file_index);
+      ok = false;
+      break;
+    }
+
     if (file_index != last_file_index) {
-      jcr->JobFiles = file_index;
       last_file_index = file_index;
+      if (file_currently_processed.GetFileIndex() > 0) {
+        processed_files.push_back(std::move(file_currently_processed));
+      }
+      file_currently_processed = ProcessedFile{file_index};
     }
 
     /*
@@ -205,35 +334,57 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
      * We save the original data pointer from the record so we can restore
      * that after the loop ends.
      */
-    rec_data = dcr->rec->data;
+    rec_data = jcr->impl->dcr->rec->data;
     while ((n = BgetMsg(bs)) > 0 && !jcr->IsJobCanceled()) {
-      dcr->rec->VolSessionId = jcr->VolSessionId;
-      dcr->rec->VolSessionTime = jcr->VolSessionTime;
-      dcr->rec->FileIndex = file_index;
-      dcr->rec->Stream = stream;
-      dcr->rec->maskedStream = stream & STREAMMASK_TYPE; /* strip high bits */
-      dcr->rec->data_len = bs->message_length;
-      dcr->rec->data = bs->msg; /* use message buffer */
+      jcr->impl->dcr->rec->VolSessionId = jcr->VolSessionId;
+      jcr->impl->dcr->rec->VolSessionTime = jcr->VolSessionTime;
+      jcr->impl->dcr->rec->FileIndex = file_index;
+      jcr->impl->dcr->rec->Stream = stream;
+      jcr->impl->dcr->rec->maskedStream
+          = stream & STREAMMASK_TYPE; /* strip high bits */
+      jcr->impl->dcr->rec->data_len = bs->message_length;
+      jcr->impl->dcr->rec->data = bs->msg; /* use message buffer */
 
       Dmsg4(850, "before writ_rec FI=%d SessId=%d Strm=%s len=%d\n",
-            dcr->rec->FileIndex, dcr->rec->VolSessionId,
-            stream_to_ascii(buf1, dcr->rec->Stream, dcr->rec->FileIndex),
-            dcr->rec->data_len);
+            jcr->impl->dcr->rec->FileIndex, jcr->impl->dcr->rec->VolSessionId,
+            stream_to_ascii(buf1, jcr->impl->dcr->rec->Stream,
+                            jcr->impl->dcr->rec->FileIndex),
+            jcr->impl->dcr->rec->data_len);
 
-      ok = dcr->WriteRecord();
+      ok = jcr->impl->dcr->WriteRecord();
       if (!ok) {
         Dmsg2(90, "Got WriteBlockToDev error on device %s. %s\n",
-              dcr->dev->print_name(), dcr->dev->bstrerror());
+              jcr->impl->dcr->dev->print_name(),
+              jcr->impl->dcr->dev->bstrerror());
         break;
       }
 
-      SendAttrsToDir(jcr, dcr->rec);
+      if (current_block_number != jcr->impl->dcr->block->BlockNumber) {
+        current_block_number = jcr->impl->dcr->block->BlockNumber;
+        SaveFullyProcessedFiles(jcr, processed_files);
+      }
+
+      if (IsAttribute(jcr->impl->dcr->rec)) {
+        file_currently_processed.AddAttribute(jcr->impl->dcr->rec);
+      }
+
+      if (me->checkpoint_interval) {
+        if (jcr->impl->dcr->VolMediaId != current_volumeid) {
+          Jmsg0(jcr, M_INFO, 0, _("Volume changed, doing checkpoint:\n"));
+          DoBackupCheckpoint(jcr);
+          current_volumeid = jcr->impl->dcr->VolMediaId;
+        } else {
+          next_checkpoint_time = DoTimedCheckpoint(jcr, next_checkpoint_time,
+                                                   me->checkpoint_interval);
+        }
+      }
+
       Dmsg0(650, "Enter bnet_get\n");
     }
     Dmsg2(650, "End read loop with %s. Stat=%d\n", what, n);
 
     // Restore the original data pointer.
-    dcr->rec->data = rec_data;
+    jcr->impl->dcr->rec->data = rec_data;
 
     if (bs->IsError()) {
       if (!jcr->IsJobCanceled()) {
@@ -268,7 +419,7 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
    * if we are at the end of the tape or we got a fatal I/O error.
    */
   if (ok || dev->CanWrite()) {
-    if (!WriteSessionLabel(dcr, EOS_LABEL)) {
+    if (!WriteSessionLabel(jcr->impl->dcr, EOS_LABEL)) {
       // Print only if ok and not cancelled to avoid spurious messages
       if (ok && !jcr->IsJobCanceled()) {
         Jmsg1(jcr, M_FATAL, 0, _("Error writing end session label. ERR=%s\n"),
@@ -281,7 +432,7 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
     Dmsg0(90, "back from write_end_session_label()\n");
 
     // Flush out final partial block of this session
-    if (!dcr->WriteBlockToDevice()) {
+    if (!jcr->impl->dcr->WriteBlockToDevice()) {
       // Print only if ok and not cancelled to avoid spurious messages
       if (ok && !jcr->IsJobCanceled()) {
         Jmsg2(jcr, M_FATAL, 0, _("Fatal append error on device %s: ERR=%s\n"),
@@ -291,18 +442,22 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
       }
       jcr->setJobStatus(JS_ErrorTerminated);
       ok = false;
+    } else if (ok && !jcr->IsJobCanceled()) {
+      // Send attributes of the final partial block of the session
+      processed_files.push_back(std::move(file_currently_processed));
+      SaveFullyProcessedFiles(jcr, processed_files);
     }
   }
 
   if (!ok && !jcr->is_JobStatus(JS_Incomplete)) {
-    DiscardDataSpool(dcr);
+    DiscardDataSpool(jcr->impl->dcr);
   } else {
     // Note: if commit is OK, the device will remain blocked
-    CommitDataSpool(dcr);
+    CommitDataSpool(jcr->impl->dcr);
   }
 
   // Release the device -- and send final Vol info to DIR and unlock it.
-  ReleaseDevice(dcr);
+  ReleaseDevice(jcr->impl->dcr);
 
   /*
    * Don't use time_t for job_elapsed as time_t can be 32 or 64 bits,
@@ -326,33 +481,23 @@ bool DoAppendData(JobControlRecord* jcr, BareosSocket* bs, const char* what)
 
   Dmsg1(100, "return from DoAppendData() ok=%d\n", ok);
   return ok;
-
-bail_out:
-  jcr->setJobStatus(JS_ErrorTerminated);
-  return false;
 }
 
 // Send attributes and digest to Director for Catalog
 bool SendAttrsToDir(JobControlRecord* jcr, DeviceRecord* rec)
 {
-  if (rec->maskedStream == STREAM_UNIX_ATTRIBUTES
-      || rec->maskedStream == STREAM_UNIX_ATTRIBUTES_EX
-      || rec->maskedStream == STREAM_RESTORE_OBJECT
-      || CryptoDigestStreamType(rec->maskedStream) != CRYPTO_DIGEST_NONE) {
-    if (!jcr->impl->no_attributes) {
-      BareosSocket* dir = jcr->dir_bsock;
-      if (AreAttributesSpooled(jcr)) { dir->SetSpooling(); }
-      Dmsg0(850, "Send attributes to dir.\n");
-      if (!jcr->impl->dcr->DirUpdateFileAttributes(rec)) {
-        Jmsg(jcr, M_FATAL, 0, _("Error updating file attributes. ERR=%s\n"),
-             dir->bstrerror());
-        dir->ClearSpooling();
-        return false;
-      }
+  if (!jcr->impl->no_attributes) {
+    BareosSocket* dir = jcr->dir_bsock;
+    if (AreAttributesSpooled(jcr)) { dir->SetSpooling(); }
+    Dmsg0(850, "Send attributes to dir.\n");
+    if (!jcr->impl->dcr->DirUpdateFileAttributes(rec)) {
+      Jmsg(jcr, M_FATAL, 0, _("Error updating file attributes. ERR=%s\n"),
+           dir->bstrerror());
       dir->ClearSpooling();
+      return false;
     }
+    dir->ClearSpooling();
   }
   return true;
 }
-
 } /* namespace storagedaemon */