+++ core/src/stored/sd_device_control_record.h
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -37,6 +37,8 @@ class StorageDaemonDeviceControlRecord : public DeviceControlRecord {
   bool DirAskSysopToMountVolume(int mode) override;
   bool DirAskSysopToCreateAppendableVolume() override;
   bool DirGetVolumeInfo(enum get_vol_info_rw writing) override;
+  bool DirAskToUpdateFileList() override;
+  bool DirAskToUpdateJobRecord() override;
   DeviceControlRecord* get_new_spooling_dcr() override;
 };