+++ core/src/stored/stored_conf.h
@@ -116,6 +116,7 @@ class StorageResource
   utime_t SDConnectTimeout = {0};       /**< Timeout in seconds */
   utime_t FDConnectTimeout = {0};       /**< Timeout in seconds */
   utime_t heartbeat_interval = {0};     /**< Interval to send hb to FD */
+  utime_t checkpoint_interval = {0};    /**< Interval to save */
   utime_t client_wait = {0};            /**< Time to wait for FD to connect */
   uint32_t max_network_buffer_size = 0; /**< Max network buf size */
   bool autoxflateonreplication